Main screen redesign
