+++ app/build.gradle
@@ -60,11 +60,12 @@ dependencies {
     implementation project(':feature:artist:list')
     implementation project(':feature:playlist:list')
     implementation project(":feature:main")
-    implementation project(":feature:player")
+    implementation project(":feature:player:mini")
+    implementation project(':feature:player:screen')
     implementation project(":feature:server:create")
     implementation project(":feature:server:list")
     implementation project(":shared:player")
     implementation(libs.bundles.koin)
     implementation(libs.navigation.compose)
-    implementation(libs.androidx.activity.compose)
+    implementation(libs.bundles.compose)
 }
