+++ app/src/main/java/ru/stresh/youamp/main/ui/MainActivity.kt
@@ -3,36 +3,39 @@ package ru.stresh.youamp.main.ui
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.animation.AnimatedContentTransitionScope
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
 import androidx.navigation.NavBackStackEntry
-import androidx.navigation.NavController
+import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-import org.koin.androidx.viewmodel.ext.android.viewModel
+import org.koin.androidx.compose.koinViewModel
 import ru.stresh.youamp.core.ui.YouAmpPlayerTheme
 import ru.stresh.youamp.feature.album.ui.AlbumInfoScreen
 import ru.stresh.youamp.feature.albums.ui.AlbumsScreen
 import ru.stresh.youamp.feature.artists.ui.ArtistsScreen
 import ru.stresh.youamp.feature.main.ui.MainScreen
-import ru.stresh.youamp.feature.player.big.ui.PlayerScreen
 import ru.stresh.youamp.feature.player.mini.ui.MiniPlayer
+import ru.stresh.youamp.feature.player.screen.ui.PlayerScreen
 import ru.stresh.youamp.feature.playlists.ui.PlaylistsScreen
 import ru.stresh.youamp.feature.server.create.ui.ServerScreen
 import ru.stresh.youamp.feature.server.list.ui.ServerListScreen
 
 class MainActivity : ComponentActivity() {
-    private val viewModel: MainViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         actionBar?.hide()
@@ -40,100 +43,150 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             YouAmpPlayerTheme {
+                val viewModel: MainViewModel = koinViewModel()
+
+                val state by viewModel.state.collectAsState()
+
                 val rootNavController = rememberNavController()
-
-                LaunchedEffect(key1 = "navigation") {
-                    viewModel
-                        .navigation
-                        .onEach {
-                            rootNavController.navigate(it.destination)
-                        }
-                        .launchIn(this)
-                }
-
                 val viewModelStoreOwner = requireNotNull(LocalViewModelStoreOwner.current)
 
-                NavHost(
-                    navController = rootNavController,
-                    startDestination = "load",
-                    modifier = Modifier.fillMaxSize()
-                ) {
-                    composable("main") {
-                        MainScreen(
-                            miniPlayer = {
-                                MiniPlayer(
-                                    onClick = {
-                                        rootNavController.navigate("player")
-                                    }
-                                )
-                            },
-                            albumsScreen = {
-                                AlbumsScreen(
-                                    viewModelStoreOwner = viewModelStoreOwner,
-                                    onAlbumClick = {
-                                        rootNavController.navigate("album/$it")
-                                    }
-                                )
-                            },
-                            artistsScreen = {
-                                ArtistsScreen(
-                                    viewModelStoreOwner = viewModelStoreOwner,
-                                    onArtistClick = {
-                                    }
-                                )
-                            },
-                            playlistsScreen = {
-                                PlaylistsScreen(
-                                    viewModelStoreOwner = viewModelStoreOwner,
-                                    onPlaylistClick = {
+                when (state) {
+                    StateUi.Content -> {
+                        Content(
+                            rootNavController = rootNavController,
+                            viewModelStoreOwner = viewModelStoreOwner
+                        )
+                    }
 
-                                    }
-                                )
-                            },
-                            onProfileClick = {
-                                rootNavController.navigate("server_list")
-                            }
-                        )
-                    }
-                    composable("album/{albumId}") {
-                        ScreenWithMiniPlayer(rootNavController) {
-                            AlbumInfoScreen(
-                                onBackClick = { rootNavController.popBackStack() },
-                                id = it.requireString("albumId")
-                            )
-                        }
-                    }
-                    composable("player") {
-                        PlayerScreen(
-                            onBackClick = { rootNavController.popBackStack() }
-                        )
-                    }
-                    composable("add_server") {
+                    StateUi.CreateNewServer -> {
                         ServerScreen(
                             onBackClick = { rootNavController.popBackStack() },
                             onCloseScreen = { rootNavController.navigate("main") }
                         )
                     }
-                    composable("edit_server/{serverId}") {
-                        ServerScreen(
-                            serverId = it.requireString("serverId").toLong(),
-                            onBackClick = { rootNavController.popBackStack() },
-                            onCloseScreen = { rootNavController.popBackStack() }
-                        )
-                    }
-                    composable("server_list") {
-                        ServerListScreen(
-                            onBackClick = { rootNavController.popBackStack() },
-                            onAddServerClick = { rootNavController.navigate("add_server") },
-                            onEditServerClick = {
-                                rootNavController.navigate("edit_server/$it")
-                            }
-                        )
-                    }
-                    composable("load") {
-                        Text(text = "Loading", modifier = Modifier.fillMaxSize())
+
+                    StateUi.Progress -> {
+
                     }
                 }
+            }
+        }
+    }
+
+    @Composable
+    private fun Content(
+        rootNavController: NavHostController,
+        viewModelStoreOwner: ViewModelStoreOwner
+    ) {
+
+        NavHost(
+            navController = rootNavController,
+            startDestination = "main",
+            modifier = Modifier
+                .fillMaxSize()
+                .background(color = MaterialTheme.colorScheme.background)
+        ) {
+            composable(
+                route = "main",
+                popEnterTransition = {
+                    fadeIn()
+                },
+                popExitTransition = {
+                    fadeOut()
+                },
+            ) {
+                MainScreen(
+                    miniPlayer = {
+                        MiniPlayer(
+                            viewModelStoreOwner = viewModelStoreOwner,
+                            onClick = {
+                                rootNavController.navigate("player")
+                            }
+                        )
+                    },
+                    albumsScreen = {
+                        AlbumsScreen(
+                            viewModelStoreOwner = viewModelStoreOwner,
+                            onAlbumClick = {
+                                rootNavController.navigate("album/$it")
+                            }
+                        )
+                    },
+                    artistsScreen = {
+                        ArtistsScreen(
+                            viewModelStoreOwner = viewModelStoreOwner,
+                            onArtistClick = {
+                            }
+                        )
+                    },
+                    playlistsScreen = {
+                        PlaylistsScreen(
+                            viewModelStoreOwner = viewModelStoreOwner,
+                            onPlaylistClick = {
+
+                            }
+                        )
+                    },
+                    onProfileClick = {
+                        rootNavController.navigate("server_list")
+                    }
+                )
+            }
+            composable("album/{albumId}") {
+                ScreenWithMiniPlayer(
+                    viewModelStoreOwner = viewModelStoreOwner,
+                    onMiniPlayerClick = {
+                        rootNavController.navigate("player")
+                    }
+                ) {
+                    AlbumInfoScreen(
+                        onBackClick = { rootNavController.popBackStack() },
+                        id = it.requireString("albumId")
+                    )
+                }
+            }
+            composable(
+                route = "player",
+                popEnterTransition = {
+                    slideIntoContainer(
+                        towards = AnimatedContentTransitionScope.SlideDirection.Up
+                    )
+                },
+                popExitTransition = {
+                    slideOutOfContainer(
+                        towards = AnimatedContentTransitionScope.SlideDirection.Down
+                    )
+                },
+                enterTransition = {
+                    slideIntoContainer(
+                        towards = AnimatedContentTransitionScope.SlideDirection.Up
+                    )
+                },
+                exitTransition = {
+                    slideOutOfContainer(
+                        towards = AnimatedContentTransitionScope.SlideDirection.Down
+                    )
+                }
+            ) {
+                PlayerScreen(
+                    onBackClick = { rootNavController.popBackStack() }
+                )
+            }
+            composable("edit_server/{serverId}") {
+                ServerScreen(
+                    serverId = it.requireString("serverId").toLong(),
+                    onBackClick = { rootNavController.popBackStack() },
+                    onCloseScreen = { rootNavController.popBackStack() }
+                )
+            }
+            composable("server_list") {
+                ServerListScreen(
+                    onBackClick = { rootNavController.popBackStack() },
+                    onAddServerClick = { rootNavController.navigate("add_server") },
+                    onEditServerClick = {
+                        rootNavController.navigate("edit_server/$it")
+                    }
+                )
             }
         }
     }
@@ -141,7 +194,8 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 internal fun ScreenWithMiniPlayer(
-    rootNavController: NavController,
+    viewModelStoreOwner: ViewModelStoreOwner,
+    onMiniPlayerClick: () -> Unit,
     content: @Composable () -> Unit
 ) {
     Surface {
@@ -150,9 +204,8 @@ internal fun ScreenWithMiniPlayer(
                 content.invoke()
             }
             MiniPlayer(
-                onClick = {
-                    rootNavController.navigate("player")
-                }
+                viewModelStoreOwner = viewModelStoreOwner,
+                onClick = onMiniPlayerClick
             )
         }
     }
