+++ app/src/main/java/ru/stresh/youamp/main/ui/StateUi.kt
@@ -1,0 +1,7 @@
+package ru.stresh.youamp.main.ui
+
+internal sealed interface StateUi {
+    data object Content : StateUi
+    data object CreateNewServer : StateUi
+    data object Progress : StateUi
+}
