+++ core/ui/src/main/java/ru/stresh/youamp/core/ui/Artwork.kt
@@ -1,16 +1,15 @@
 package ru.stresh.youamp.core.ui
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.shape.CornerBasedShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Album
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import coil.compose.SubcomposeAsyncImage
@@ -25,7 +24,7 @@ private fun ArtworkPreview() {
 fun Artwork(
     artworkUrl: String?,
     modifier: Modifier = Modifier,
-    shape: CornerBasedShape = MaterialTheme.shapes.medium
+    shape: CornerBasedShape = MaterialTheme.shapes.large
 ) {
     SubcomposeAsyncImage(
         model = artworkUrl,
@@ -45,11 +44,10 @@ fun Artwork(
 
 @Composable
 fun ArtworkPlaceholder() {
-    Image(
+    Icon(
         imageVector = Icons.Rounded.Album,
         contentDescription = "placeholder",
-        modifier = Modifier
-            .alpha(0.3f)
-            .background(MaterialTheme.colorScheme.onSecondaryContainer)
+        tint = MaterialTheme.colorScheme.onSecondary,
+        modifier = Modifier.background(MaterialTheme.colorScheme.secondary)
     )
 }
