+++ feature/player/mini/.gitignore
