+++ feature/player/mini/proguard-rules.pro
