+++ feature/server/create/src/main/java/ru/stresh/youamp/feature/server/create/Di.kt
@@ -8,7 +8,7 @@ import ru.stresh.youamp.feature.server.create.ui.ServerCreateViewModel
 
 val serverCreateModule = module {
     factory<ServerRepository> { ServerRepositoryImpl(get()) }
-    viewModel { (serverId: Long) ->
+    viewModel { (serverId: Long?) ->
         ServerCreateViewModel(serverId, get())
     }
 }
