+++ settings.gradle
@@ -22,7 +22,8 @@ include ':shared:player'
 include ':feature:album:list'
 include ':feature:album:info'
 include ':feature:main'
-include ':feature:player'
+include ':feature:player:mini'
+include ':feature:player:screen'
 include ':feature:server:create'
 include ':feature:server:list'
 include ':feature:artist:list'
