--- shared/player/src/main/java/ru/stersh/jcsplayer/shared/player/progress/PlayerProgress.kt
@@ -1,8 +1,0 @@
-package ru.stersh.youamp.shared.player.progress
-
-data class PlayerProgress(
-    val currentTimeMs: Long,
-    val totalTimeMs: Long,
-    val currentTime: String,
-    val totalTime: String,
-)
