+++ shared/player/src/main/java/ru/stersh/youamp/shared/player/progress/PlayerProgress.kt
@@ -1,0 +1,11 @@
+package ru.stersh.youamp.shared.player.progress
+
+data class PlayerProgress(
+    val currentTimeMs: Long,
+    val totalTimeMs: Long,
+    val currentTime: String,
+    val totalTime: String,
+) {
+    val progressPercent: Float
+        get() = currentTimeMs.toFloat() / totalTimeMs
+}
